+++ src/platform/win32_thread.cc
@@ -3,27 +3,21 @@
 namespace platform
 {
 
-DWORD WINAPI
-Win32ThreadFunc(LPVOID lpParam)
-{
+DWORD WINAPI Win32ThreadFunc(LPVOID lpParam) {
   Thread* ti = (Thread*)lpParam;
   u64 ret = ti->func(ti->arg);
   ti->return_value = ret;
-  return ret;
+  return (DWORD)ret;
 }
 
 
-u64
-ThreadId()
-{
+u64 ThreadId() {
   // TODO: Replace with intrinsic __readgsqword
   // https://docs.microsoft.com/en-us/cpp/intrinsics/readgsbyte-readgsdword-readgsqword-readgsword?view=vs-2019
   return (u64)GetCurrentThreadId();
 }
 
-b8
-ThreadCreate(Thread* t)
-{
+b8 ThreadCreate(Thread* t) {
   if (t->id) return false;
   t->handle = CreateThread(
       nullptr,          // Default security attributes.
@@ -35,31 +29,23 @@ ThreadCreate(Thread* t)
   return t->handle != nullptr;
 }
 
-void
-ThreadYield()
-{
+void ThreadYield() {
   SwitchToThread();
 }
 
-b8
-ThreadJoin(Thread* t)
-{
+b8 ThreadJoin(Thread* t) {
   WaitForSingleObject(t->handle, INFINITE);
   GetExitCodeThread(t->handle, (LPDWORD)&t->return_value);
   CloseHandle(t->handle);
   return 0;
 }
 
-void
-ThreadExit(Thread* t, u64 value)
-{
+void ThreadExit(Thread* t, u64 value) {
   t->return_value = value;
   ExitThread((DWORD)t->return_value);
 }
 
-b8
-MutexCreate(Mutex* m)
-{
+b8 MutexCreate(Mutex* m) {
   m->handle = CreateMutex(
       nullptr,      // Default security attributes.
       false,        // Not owned initially.
@@ -73,25 +59,19 @@ MutexCreate(Mutex* m)
   return true;
 }
 
-void
-MutexLock(Mutex* m)
-{
+void MutexLock(Mutex* m) {
   if (WaitForSingleObject(m->handle, INFINITE) == WAIT_FAILED) {
     printf("mutex_lock error: %d\n", GetLastError());
   }
 }
 
-void
-MutexUnlock(Mutex* m)
-{
+void MutexUnlock(Mutex* m) {
   if (!ReleaseMutex(m->handle)) {
     printf("mutex_unlock error: %d\n", GetLastError());
   }
 }
 
-void
-MutexFree(Mutex* m)
-{
+void MutexFree(Mutex* m) {
   CloseHandle(m->handle);
 }
